tests/integration/apps/grpc-app/requirements.txt
starlette
uvicorn
protobuf==5.29.6
grpcio==1.68.1
grpcio-tools==1.68.1
